merged: changes from trunk. revs: r27731:27881
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/pvr-testing2@27883 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/lib/libRTMP/AMFObject.cpp
@@ -321,7 +321,7 @@ int RTMP_LIB::AMFObject::Encode(char * pBuffer, int nSize) const
     int nRes = m_properties[i].Encode(pBuffer, nSize);
     if (nRes == -1)
     {
-      CLog::Log(LOGERROR,"AMFObject::Encode - failed to encode property in index %d", i);
+      CLog::Log(LOGERROR,"AMFObject::Encode - failed to encode property in index %d", (int)i);
     }
     else
     {
